Added gradient backgrounds to interior pages
Added the gradient backgrounds and also had to make sure my source code
matched up with that which was in the actual lesson so I pretty much
copied and pasted the source code to make sure I was on the right track.
For some reason that spacing between the header and the hero container
were not spacing correctly but after I copied and pasted all of the
source code, it did what it was supposed to do. I'll be sure to be more
careful of my code even though it is not clear at all times what they
are asking for.